v3.18.0 exchange rate and account models (#33) [release-type final]
* v3.18.0-SNAPSHOT exchange rate and account models 

### New Features
* Added new models:
  - `ExchangeRate` to describe the exchange rate between two currencies.
  - `Account` to describe more varied accounts in a consistent manner.
This has the following sub-types defined:
    - `BankAccount`
    - `IbanAccount`
    - `IfsAccount`
    - `SwiftAccount`
    - `MobileWalletAccount`
* The limitations on the `id` field of the `Institution` model have
been removed to make the field suitable for a wider range of
applications. The values in this field need no longer be defined by
Electrum and similar values as recognised at third parties may now be
carried in this field. However, care should be taken during
implementation to ensure that different parties agree on the values
which will be present in these fields.

* review comments

* minor typo

* trigger circle

* mask account ID

* finalise version and release notes